requirements.txt
greenlet==3.0.1
psycopg2-binary==2.9.9
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
shiboken6==6.6.0
SQLAlchemy==1.4.11
SQLAlchemy-Utils==0.41.1
